Switch to pessimistic SQL disconnect handling
Previously we were using optimistic disconnect handling for the database
connection pool: assume that all the connections are live and recycle
them before a certain timeout just in case the server kills the
connections after a certain period of inactivity. This approach reduced
the frequency of "MySQL went away" errors, but it turns out that it
didn't eliminate them.

As such, this change switches to pessimistic disconnect handling: before
using a given connection, it is first tested for liveness by SQLAlchemy.
This is implemented by passing `pool_pre_ping` to the `create_engine`
call [1]. Note that this option was only introduced in SQLAlchemy 1.2
[2] so this change also updates the SQLAlchemy dependency to the latest
release.

[1] http://docs.sqlalchemy.org/en/latest/core/pooling.html#pool-disconnects-pessimistic
[2] https://twitter.com/sqlalchemy/status/884415051125452802